+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Layer with isNeverVisibleInTree will not be initialized in menu tree.
 - Issue #621: WMTS support for overviewMap implemented.
 - Update print configuration for tiledWMS, layers are printed as tiledwms with tileSize.
+- WFSSearch, has been adjusted so that it zooms directly to the first feature found if no `resultList` is configured.
 
 ### Deprecated